Updates multiboot code for amd64
The old version was 64-bit incompatible (assumed longs were 32 bits).
The FSF has a newer version anyway.

Note our old style of determining which memory to use only works for the
first 3GB or so of pmem (at least on my test machines).

+++ kern/include/ros/common.h
@@ -12,8 +12,8 @@
 #include <stdbool.h>
 
 typedef uintptr_t physaddr_t;
-typedef ssize_t intreg_t;
-typedef size_t uintreg_t;
+typedef long intreg_t;
+typedef unsigned long uintreg_t;
 
 #ifndef NULL
 #define NULL ((void*) 0)
